extracted more dto's
