+++ rest-api/src/main/java/be/xplore/notifyme/controller/OrganisationManagerController.java
@@ -3,7 +3,7 @@ package be.xplore.notifyme.controller;
 import be.xplore.notifyme.domain.TeamApplication;
 import be.xplore.notifyme.domain.TeamApplicationKey;
 import be.xplore.notifyme.dto.organisationapplication.ApplicationUsernameDto;
-import be.xplore.notifyme.dto.OrganisationDto;
+import be.xplore.notifyme.dto.organisation.OrganisationDto;
 import be.xplore.notifyme.dto.OrganisationLimitedInfoDto;
 import be.xplore.notifyme.dto.UserApplicationResponseDto;
 import be.xplore.notifyme.services.ITeamApplicationService;
