+++ rest-api/src/main/java/be/xplore/notifyme/controller/UserController.java
@@ -5,7 +5,7 @@ import be.xplore.notifyme.domain.Line;
 import be.xplore.notifyme.domain.Team;
 import be.xplore.notifyme.domain.TeamApplication;
 import be.xplore.notifyme.dto.organisationapplication.ApplicationOrgNameDto;
-import be.xplore.notifyme.dto.NotificationDto;
+import be.xplore.notifyme.dto.notification.NotificationDto;
 import be.xplore.notifyme.dto.OrganisationsLimitedInfoDto;
 import be.xplore.notifyme.dto.UserRegistrationDto;
 import be.xplore.notifyme.dto.communicationpreference.GetCommunicationPreferenceDto;
