+++ rest-api/src/main/java/be/xplore/notifyme/dto/organisation/OrganisationDto.java
@@ -1,4 +1,4 @@
-package be.xplore.notifyme.dto;
+package be.xplore.notifyme.dto.organisation;
 
 import be.xplore.notifyme.domain.Organisation;
 import be.xplore.notifyme.domain.OrganisationUser;
